+++ let-plugin/src/main/groovy/com/canelmas/let/LetPlugin.groovy
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2017 Can Elmas
+ * Copyright (C) 2018 Can Elmas
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +17,8 @@
 package com.canelmas.let
 
 import com.android.build.gradle.AppPlugin
+import com.android.build.gradle.FeaturePlugin
+import com.android.build.gradle.InstantAppPlugin
 import com.android.build.gradle.LibraryPlugin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -26,23 +28,26 @@ class LetPlugin implements Plugin<Project> {
     @Override
     void apply(Project project) {
 
-        makeSureProjectIsAndroidAppOrLib(project)
+        makeSureProjectIsCompatible(project)
 
         project.android.registerTransform(new LetTransform(project))
 
         project.dependencies {
             implementation 'org.aspectj:aspectjrt:1.8.9'
-            implementation 'com.canelmas.let:let-runtime:1.0.0-beta1'
-            implementation 'com.canelmas.let:let-annotations:1.0.0-beta1'
+            implementation 'com.canelmas.let:let-runtime:1.0.0-beta2'
+            implementation 'com.canelmas.let:let-annotations:1.0.0-beta2'
         }
     }
 
-    private void makeSureProjectIsAndroidAppOrLib(Project project) {
-        def hasApp = project.plugins.withType(AppPlugin)
-        def hasLib = project.plugins.withType(LibraryPlugin)
+    private void makeSureProjectIsCompatible(Project project) {
+        def hasAppPlugin = project.plugins.withType(AppPlugin)
+        def hasLibPlugin = project.plugins.withType(LibraryPlugin)
+        def hasFeaturePlugin = project.plugins.withType(FeaturePlugin)
+        def hasInstantAppPlugin = project.plugins.withType(InstantAppPlugin)
 
-        if (!hasApp && !hasLib) {
-            throw new IllegalStateException("'android' or 'android-library' plugin required.")
+        if (!hasAppPlugin && !hasLibPlugin && !hasFeaturePlugin && !hasInstantAppPlugin) {
+            throw new IllegalStateException("Project must have one the following plugins applied " +
+                    ": 'android', 'android-library', 'feature', 'instantapp'")
         }
     }