+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+Version 1.0.0-beta2 *(2018-02-11)*
+----------------------------------
+* Add `FeaturePlugin` and `InstantAppPlugin` support [#28](https://github.com/canelmas/let/issues/28)
+* Rename `Delayed Tasks` to `ScheduledTasks`
+* Update copyright
+
 Version 1.0.0-beta1 *(2017-09-04)*
 ----------------------------------
 * Use Transform API for weaving [#23](https://github.com/canelmas/let/issues/23)
@@ -18,5 +24,4 @@ Version 0.1.10 *(2016-03-20)*
 
 Version 0.1.9 *(2015-11-12)*
 ----------------------------
-
-initial release
+* initial release